rename: myroom 폴더명 수정
