+++ project04/frontend/src/components/content/canvas/maps/structures/myRoom/elements/MyRoomRightWall.jsx
